- fix read notification infinite loop - move notifications.ts to /engine/
