+++ src/app/MenuDesktop.svelte
@@ -24,7 +24,7 @@
   import {slowConnections} from "src/app/state"
   import {router} from "src/app/util/router"
   import {env, hasNewMessages} from "src/engine"
-  import {hasNewNotifications} from "src/domain/notifications"
+  import {hasNewNotifications} from "src/engine/notifications"
 
   const {page} = router
