+++ src/app/views/ChannelsList.svelte
@@ -12,7 +12,7 @@
   import ChannelsListItem from "src/app/views/ChannelsListItem.svelte"
   import {router} from "src/app/util/router"
   import {channels, hasNewMessages} from "src/engine"
-  import {setChecked} from "src/domain/notifications"
+  import {setChecked} from "src/engine/notifications"
 
   const activeTab = window.location.pathname.slice(1) === "channels" ? "conversations" : "requests"
   const setActiveTab = tab => {
