+++ src/app/views/Onboarding.svelte
@@ -19,7 +19,7 @@
     broadcastUserData,
   } from "src/engine"
   import {router} from "src/app/util/router"
-  import {setChecked} from "src/domain/notifications"
+  import {setChecked} from "src/engine/notifications"
 
   export let stage = "intro"
   export let invite = null
